+++ plugin/src/App/StatusPages/Settings/init.lua
@@ -232,9 +232,9 @@ function SettingsPage:render()
 			CheckForUpdates = e(Setting, {
 				id = "checkForUpdates",
 				name = "Check For Updates",
-				description = "Notify about newer compatible Rojo releases",
+				description = "Notify about newer compatible Roxo releases",
 				locked = Version.isApiBlocked(),
-				lockedTooltip = "(HTTP requests to api.github.com are blocked, Rojo cannot fetch what the latest version is.)",
+				lockedTooltip = "(HTTP requests to api.github.com are blocked, Roxo cannot fetch what the latest version is.)",
 				transparency = self.props.transparency,
 				layoutOrder = layoutIncrement(),
 			}),
@@ -253,7 +253,7 @@ function SettingsPage:render()
 			AutoConnectPlaytestServer = e(Setting, {
 				id = "autoConnectPlaytestServer",
 				name = "Auto Connect Playtest Server",
-				description = "Automatically connect game server to Rojo when playtesting while connected in Edit",
+				description = "Automatically connect game server to Roxo when playtesting while connected in Edit",
 				tag = "unstable",
 				transparency = self.props.transparency,
 				layoutOrder = layoutIncrement(),
@@ -308,7 +308,7 @@ function SettingsPage:render()
 			TimingLogsEnabled = e(Setting, {
 				id = "timingLogsEnabled",
 				name = "Timing Logs",
-				description = "Toggle logging timing of internal actions for benchmarking Rojo performance",
+				description = "Toggle logging timing of internal actions for benchmarking Roxo performance",
 				tag = "debug",
 				transparency = self.props.transparency,
 				layoutOrder = layoutIncrement(),